+++ packages/auth/src/lib/auth-form/auth-microsoftb2c.component.ts
@@ -6,7 +6,8 @@ import {
   EventEmitter,
   Inject
 } from '@angular/core';
-import { MsalBroadcastService, MsalService, MSAL_GUARD_CONFIG} from '@azure/msal-angular';
+
+import { MSAL_GUARD_CONFIG } from '@azure/msal-angular';
 import {
   InteractionStatus,
   AuthenticationResult,
@@ -16,42 +17,45 @@ import {
   InteractionRequiredAuthError
 } from '@azure/msal-browser';
 import { ConfigService } from '@igo2/core';
-import { AuthMicrosoftOptions, MSPMsalGuardConfiguration } from '../shared/auth.interface';
+import { AuthMicrosoftb2cOptions, MSPMsalGuardConfiguration } from '../shared/auth.interface';
 import { AuthService } from '../shared/auth.service';
 import { filter, takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
+import { MsalBroadcastServiceb2c } from '../shared/auth-msalBroadcastServiceb2c.service';
+import { MsalServiceb2c } from '../shared/auth-msalServiceb2c.service.';
 
 @Component({
-  selector: 'igo-auth-microsoft',
-  templateUrl: './auth-microsoft.component.html',
-  styleUrls: ['./auth-microsoft.component.scss'],
+  selector: 'igo-auth-microsoftb2c',
+  templateUrl: './auth-microsoftb2c.component.html',
+  styleUrls: ['./auth-microsoftb2c.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class AuthMicrosoftComponent {
-  private options: AuthMicrosoftOptions;
+export class AuthMicrosoftb2cComponent {
+  private options: AuthMicrosoftb2cOptions;
   private readonly _destroying$ = new Subject<void>();
   @Output() login: EventEmitter<boolean> = new EventEmitter<boolean>();
-  private broadcastService: MsalBroadcastService;
+  private broadcastService: MsalBroadcastServiceb2c;
 
   constructor(
     private authService: AuthService,
     private config: ConfigService,
     private appRef: ApplicationRef,
-    private msalService: MsalService,
+    private msalService: MsalServiceb2c,
     @Inject(MSAL_GUARD_CONFIG) private msalGuardConfig: MSPMsalGuardConfiguration[],
   ) {
-    this.options = this.config.getConfig('auth.microsoft') || {};
+
+    this.options = this.config.getConfig('auth.microsoftb2c') || {};
 
     this.msalService.instance = new PublicClientApplication({
-      auth: this.options,
+      auth: this.options.browserAuthOptions,
       cache: {
         cacheLocation: 'sessionStorage'
       }
     });
 
-    this.broadcastService = new MsalBroadcastService(this.msalService.instance, this.msalService);
+    this.broadcastService = new MsalBroadcastServiceb2c(this.msalService.instance, this.msalService);
 
-    if (this.options.clientId) {
+    if (this.options.browserAuthOptions.clientId) {
       this.broadcastService.inProgress$
       .pipe(
         filter((status: InteractionStatus) => status === InteractionStatus.None),
@@ -66,7 +70,7 @@ export class AuthMicrosoftComponent {
     }
   }
 
-  public loginMicrosoft() {
+  public loginMicrosoftb2c() {
     this.msalService.loginPopup({...this.getConf().authRequest} as PopupRequest)
     .subscribe((response: AuthenticationResult) => {
       this.msalService.instance.setActiveAccount(response.account);
@@ -78,9 +82,8 @@ export class AuthMicrosoftComponent {
     this.msalService.instance
       .acquireTokenSilent(this.getConf().authRequest as SilentRequest)
       .then((response: AuthenticationResult) => {
-        const tokenAccess = response.accessToken;
-        const tokenId = response.idToken;
-        this.authService.loginWithToken(tokenAccess, 'microsoft', { tokenId } ).subscribe(() => {
+        const token = response.idToken;
+        this.authService.loginWithToken(token, 'microsoftb2c').subscribe(() => {
           this.appRef.tick();
           this.login.emit(true);
         });
@@ -90,13 +93,12 @@ export class AuthMicrosoftComponent {
           // fallback to interaction when silent call fails
           return this.msalService.acquireTokenPopup(this.getConf().authRequest as SilentRequest);
         }
-        console.log(error);
-      }).catch(error => {
-        console.log('Silent token fails');
-      });
+        }).catch(error => {
+          console.log('Silent token fails');
+        });
   }
 
   private getConf(): MSPMsalGuardConfiguration {
-    return this.msalGuardConfig.filter(conf => conf.type === 'add')[0];
+    return this.msalGuardConfig.filter(conf => conf.type === 'b2c')[0];
   }
 }