+++ packages/auth/src/lib/shared/auth.interface.ts
@@ -1,3 +1,6 @@
+import { BrowserAuthOptions } from '@azure/msal-browser';
+import { MsalGuardConfiguration } from '@azure/msal-angular';
+
 export interface AuthInternOptions {
   enabled?: boolean;
 }
@@ -16,6 +19,37 @@ export interface AuthGoogleOptions {
 export interface AuthMicrosoftOptions {
   enabled?: boolean;
   clientId: string;
+  redirectUri?: string;
+  authority?: string;
+}
+
+export interface AuthMicrosoftb2cOptions {
+  enabled?: boolean;
+  browserAuthOptions?: BrowserAuthOptions;
+  options?: AuthMicrosoftb2cOptionsOptions;
+  scopes?: string[];
+}
+
+export interface AuthMicrosoftb2cOptionsOptions {
+  names: AuthMicrosoftb2cOptionsNames;
+  authorities: AuthMicrosoftb2cOptionsAuthorities;
+  authorityDomain: string[];
+}
+
+export interface AuthMicrosoftb2cOptionsNames {
+  signUpSignIn: string;
+  forgotPassword: string;
+  editProfile: string;
+}
+
+export interface AuthMicrosoftb2cOptionsAuthorities {
+  signUpSignIn: AuthMicrosoftb2cOptionsAuthority;
+  forgotPassword: AuthMicrosoftb2cOptionsAuthority;
+  editProfile: AuthMicrosoftb2cOptionsAuthority;
+}
+
+export interface AuthMicrosoftb2cOptionsAuthority {
+  authority: string;
 }
 
 export interface AuthOptions {
@@ -29,6 +63,7 @@ export interface AuthOptions {
   facebook?: AuthFacebookOptions;
   google?: AuthGoogleOptions;
   microsoft?: AuthMicrosoftOptions;
+  microsoftb2c?: AuthMicrosoftb2cOptions;
   trustHosts?: string[];
   profilsGuard?: string[];
   hostsWithCredentials?: WithCredentialsOptions[];
@@ -37,6 +72,26 @@ export interface WithCredentialsOptions {
   withCredentials?: boolean;
   domainRegFilters?: string;
  }
+
+export interface MSPMsalGuardConfiguration extends MsalGuardConfiguration {
+  type: string;
+}
+
+export interface MSPMsalGuardConfiguration extends MsalGuardConfiguration {
+  type: string;
+}
+
+export interface MSPMsalGuardConfiguration extends MsalGuardConfiguration {
+  type: string;
+}
+
+export interface MSPMsalGuardConfiguration extends MsalGuardConfiguration {
+  type: string;
+}
+
+export interface MSPMsalGuardConfiguration extends MsalGuardConfiguration {
+  type: string;
+}
 
 export interface User {
   source?: string;
@@ -48,4 +103,8 @@ export interface User {
   isExpired?: boolean;
   admin?: boolean;
   defaultContextId?: string;
+}
+
+export interface IInfosUser {
+  tokenId: string;
 }