+++ packages/geo/src/lib/datasource/shared/datasources/wms-datasource.ts
@@ -113,7 +113,8 @@ export class WMSDataSource extends DataSource {
         'wms'
       );
     } else {
-      initOgcFilters.advancedOgcFilters = (initOgcFilters.pushButtons || initOgcFilters.checkboxes || initOgcFilters.radioButtons)
+      initOgcFilters.advancedOgcFilters = (initOgcFilters.pushButtons || initOgcFilters.checkboxes
+        || initOgcFilters.radioButtons || initOgcFilters.select)
         ? false
         : true;
       if (initOgcFilters.pushButtons){
@@ -124,6 +125,9 @@ export class WMSDataSource extends DataSource {
       }
       if (initOgcFilters.radioButtons){
         initOgcFilters.radioButtons.selectorType = 'radioButton';
+      }
+      if (initOgcFilters.select){
+        initOgcFilters.select.selectorType = 'select';
       }
     }