+++ packages/geo/src/lib/search/shared/search.interfaces.ts
@@ -1,5 +1,6 @@
 import { Observable } from 'rxjs';
 
+import { CommonVectorStyleOptions } from '../../utils/commonVectorStyle.interface';
 import { SearchSource } from './sources/source';
 
 export interface Research {
@@ -19,5 +20,10 @@ export interface SearchResult<T = { [key: string]: any }> {
     icon: string;
     score?: number;
     nextPage?: boolean;
+  };
+  style?: {
+    base?: CommonVectorStyleOptions;
+    selection?: CommonVectorStyleOptions;
+    focus?: CommonVectorStyleOptions;
   };
 }